+++ solo-chains/runtime/dancelight/src/lib.rs
@@ -1741,11 +1741,7 @@ construct_runtime! {
         XcmPallet: pallet_xcm = 90,
 
         // Bridging stuff
-        // https://github.com/paritytech/polkadot-sdk/blob/2ae79be8e028a995b850621ee55f46c041eceefe/cumulus/parachains/runtimes/bridge-hubs/bridge-hub-westend/src/lib.rs#L560C1-L560C64
-        //EthereumInboundQueue: snowbridge_pallet_inbound_queue = 80,
         EthereumOutboundQueue: snowbridge_pallet_outbound_queue = 101,
-        // TODO: already exists, at index 243
-        //EthereumBeaconClient: snowbridge_pallet_ethereum_client = 82,
         EthereumSystem: snowbridge_pallet_system = 103,
 
         // Migration stuff
